requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
--find-links https://prototyp3-dev.github.io/pip-wheels-riscv/wheels/

absl-py==2.1.0
    # via
    #   keras
    #   tensorboard
    #   tensorflow
astunparse==1.6.3
    # via tensorflow
certifi==2023.11.17
    # via requests
charset-normalizer==3.3.2
    # via requests
cytoolz==0.12.2
    # via eth-utils
eth-abi==4.2.1
    # via python-cartesi
eth-hash==0.5.2
    # via eth-utils
eth-typing==3.5.2
    # via
    #   eth-abi
    #   eth-utils
eth-utils==2.3.1
    # via eth-abi
exceptiongroup==1.2.0
    # via pytest
flatbuffers==24.3.25
    # via tensorflow
gast==0.5.4
    # via tensorflow
google-pasta==0.2.0
    # via tensorflow
grpcio==1.64.1
    # via
    #   tensorboard
    #   tensorflow
h5py==3.11.0
    # via
    #   keras
    #   tensorflow
idna==3.6
    # via requests
iniconfig==2.0.0
    # via pytest
keras==3.3.3
    # via
    #   mtcnn
    #   tensorflow
keras-facenet==0.3.2
    # via -r requirements.in
libclang==18.1.1
    # via tensorflow
markdown==3.6
    # via tensorboard
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via werkzeug
mdurl==0.1.2
    # via markdown-it-py
ml-dtypes==0.3.2
    # via
    #   keras
    #   tensorflow
mtcnn==0.1.1
    # via keras-facenet
namex==0.0.8
    # via keras
numpy==1.26.4
    # via
    #   -r requirements.in
    #   h5py
    #   keras
    #   ml-dtypes
    #   opencv-python
    #   opt-einsum
    #   scipy
    #   tensorboard
    #   tensorflow
opencv-python==4.10.0.82
    # via mtcnn
opt-einsum==3.3.0
    # via tensorflow
optree==0.11.0
    # via keras
packaging==23.2
    # via
    #   pytest
    #   tensorflow
parsimonious==0.9.0
    # via eth-abi
pluggy==1.3.0
    # via pytest
protobuf==4.25.3
    # via
    #   tensorboard
    #   tensorflow
pycryptodome==3.19.0
    # via
    #   -r requirements.in
    #   python-cartesi
pydantic==1.10.13
    # via python-cartesi
pygments==2.18.0
    # via rich
pytest==7.4.4
    # via
    #   -r requirements.in
    #   python-cartesi
python-cartesi==0.1.0
    # via -r requirements.in
regex==2023.12.25
    # via parsimonious
requests==2.31.0
    # via
    #   -r requirements.in
    #   python-cartesi
    #   tensorflow
rich==13.7.1
    # via keras
scipy==1.10.1
    # via -r requirements.in
six==1.16.0
    # via
    #   astunparse
    #   google-pasta
    #   tensorboard
    #   tensorflow
tensorboard==2.16.2
    # via tensorflow
tensorboard-data-server==0.7.2
    # via tensorboard
tensorflow==2.16.1
    # via -r requirements.in
tensorflow-io-gcs-filesystem==0.37.0
    # via tensorflow
termcolor==2.4.0
    # via tensorflow
tomli==2.0.1
    # via pytest
toolz==0.12.0
    # via cytoolz
typing-extensions==4.9.0
    # via
    #   eth-typing
    #   optree
    #   pydantic
    #   tensorflow
urllib3==2.1.0
    # via requests
werkzeug==3.0.3
    # via tensorboard
wheel==0.43.0
    # via astunparse
wrapt==1.16.0
    # via tensorflow

# The following packages are considered to be unsafe in a requirements file:
# setuptools
